requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
instagrapi>=2.0.0
huggingface_hub>=1.0.0
google-genai>=1.0.0
replicate>=1.0.0
